improve XML validation during CI - reorg to put the validation info and scripts in a subfolder - add XSD schema files and do some partial validation on the UDL and autoCompletion files     (enough to verify it's UDL2.0/2.1 syntax and to weed out non-UDL files)
(These were tested/developed in an experimental branch, and verified against the old mix
of UDL1 and UDL2 files, and then again tested against the updated list with all UDL2 files from #149.
Developing these validations is actually what caught the couple of XML files that weren't
valid UDL that I recently fixed or deleted)
