+++ .validators/autocomplete.xsd
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
+
+<xs:element name="NotepadPlus">
+  <xs:complexType>
+    <xs:sequence>
+      <xs:element name="AutoComplete" maxOccurs="unbounded"/>
+    </xs:sequence>
+  </xs:complexType>
+</xs:element>
+
+</xs:schema>
